kaikki paitsi 4.22

+++ osa4/blogilista/src/controllers/blog.js
@@ -1,8 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => { 
-	const blogs = await Blog.find({})
+	const blogs = await Blog.find({}).populate('adder')
+
 	response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -17,12 +21,20 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
 	const body = request.body
+	const token = request.token
+
+	const decodedToken = jwt.verify(token, process.env.SECRET)
+	if (!token || !decodedToken.id) {
+	  return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
 
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
 		url: body.url,
-		likes: body.likes || 0
+		likes: body.likes || 0,
+		adder: user._id
 	})
 
 	const savedBlog = await blog.save()
@@ -30,8 +42,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id)
-	response.status(204).end()
+	const token = request.token
+
+	const decodedToken = jwt.verify(token, process.env.SECRET)
+	if (!token || !decodedToken.id) {
+	  return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
+
+	const blog = await Blog.findById(request.params.id).populate("adder")
+
+	if ( blog.adder._id.toString() === user._id.toString() ){
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	} else {
+		response.status(401).json({ error: "you're not the owner of this blog" }).end()
+	}
 })
 
 blogsRouter.put('/:id', async (request, response) => {
